+++ service/app/middleware/Cors.php
@@ -21,7 +21,7 @@ class Cors implements MiddlewareInterface
             return response('', 204, [
                 'Access-Control-Allow-Origin'      => $allowedOrigin,
                 'Access-Control-Allow-Methods'     => 'GET,POST,PUT,DELETE,OPTIONS',
-                'Access-Control-Allow-Headers'     => 'Authorization,Content-Type,API-Version',
+                'Access-Control-Allow-Headers'     => 'Authorization,Content-Type',
                 'Access-Control-Max-Age'           => '86400',
             ]);
         }